Remove manifest and deploy script.
I wasn't really using them to update the website. We need a better
process if we are going to be automated about the whole affair. I
think there isn't a compelling reason to set up something like
that yet, though, so for now manifest.txt and deploy.sh go away.