Add README; limit network overhead; time callbacks
It turns out that i was misunderstanding the effective use of the
dcc.Store component. Since all the callbacks are computed server-side,
there is little to be gained by storing the JSONified dataframe in
the client's browser, since it has to be passed back over the network
in order run the desired computations. Caching server-side is the
proper solution to this issue. All that is now stored in the browser
is a signal that a given plant/unit has been read from SQL and is
ready to have plots generated for it. This minimal amount of data
(likely on the order of <10B) is practically free.

As a note for the future: Right now, reading the data from SQL is
a major pain point and adds about 5s of overhead to every single
query that isn't cached. The best way to deal with this would be to
simply make SQL reads faster, but it would also be possible to copy
a series of flat serialized JSON or CSV files to a location on disk
located reasonably close to the server. As a ballpark estimate, i
was getting that the data for a given plant/unit over 18 years was
somewhere between 5MB and 10MB compressed on disk; since we've got
about 650 plant/units, that would end up taking up somewhere between
3.2GB and 6.5GB, which is not trivial but also not unreasonable.