Add fancy new rezoomability to emis time series.
The new functionality is to show "trend lines", the sort of
line + area plot that showed up on the map page for gload,
whenever we are zoomed out to the weekly or daily aggregation
level. This visually preserves some of the variability of the data,
but allows very fast rendering. Zooming in farther loads the actual
hourly data points.