Add scattergl plots for emissions intensity vs CF.
This also does a better job of documenting the development process
and the tools i use to debug things locally. I still need to come
up with a snazzier solution to static site generation, and the
remaining piece of un-reproduceability is the fact that the gzipped
CSVs that the unit-level dashboard relies on do not live inside the
git repository.