Speed up quantile computation.
Prior to this change, we were computing histograms in an inefficient
manner. Because we know the data is sorted by time on the way in,
we only need to check whether we've crossed a date boundary (e.g.
next day, next week, next month) to start a new bin rather than
calling timeInterval#floor on every Date object, which is somewhat
expensive. This speeds things up considerably, although ScatterGL
is still eating up tons of time...