First pass at histograms
I've decided that the EI vs CF scatter plot needs an overhaul. For
now i'm just limiting the number of points to 2000 so that
rendering doesn't take eons, and using 'scatter' rather than
'scattergl' so that we can depend on only one Plotly package: the
Cartesian bundle (plotly-cartesian-latest). Things are still pretty
slow, which i'm not thrilled about. Also, i need to make it possible
to select which gas to look at for the EI histogram.